Add testing utilities (#9)
* Add Google.Mock

This module makes it much easier to ensure that correct values are being
returned when using Mox to mock an adapter. This works well as a stub or
even just a function to call to return the proper value

* Add Adapter.test_message/2 function

This is a wrapper around Broadway.test_message/2. It's use to allow us
to create messages and send them through the pipeline. It basically can
use the configured adapter to go from a GenesisPubSub.Message to a
Broadway.Message

* Add testing guide

Testing muliprocess consumers was a little tricky so I wanted to add a
guide that can walk through the challenges and possible solutions to
using Mox in Broadway pipeline.

* Bump minor version

Co-authored-by: Stephen Bolton <skbolton@users.noreply.github.com>