Add support for `test_batch/2` (#16)
* Refactor `test_message/2` to use `push_message/2`
* Add `test_batch/2` to Consumer
* Remove `test_message/2` from adapter callbacks
* Add `pack/2` adapter callback

To test sending several messages through a pipeline Broadway has a
`test_batch/3` function. It takes in the broadway module to send to,
data to wrap, and some options. The options can contain a metadata key
to supply to all messages. The issue is that we would need to pass a
unique metadata to each message being sent through. According to this
[issue](https://github.com/dashbitco/broadway/issues/197) the solution
is to use the `push_messages/2` function and manually build a Broadway
Message instead of using the testing wrappers.

We already have the `test_message/2` function for sending through a
single message and it wrapped the broadway provided `test_message`
function. I decided it would be better off to use the `push_messages/2`
function for both the single message and batch message calls for
consistency.

This lead me to refactor `test_message` along with writing `test_batch`
to use `push_messages`. As I was doing so I realized how much work we
were doing in the Adapter to send these test messages. This meant that
for each adapter we might implement they were going to have to do some
copy and pasting. Because of this I added the `pack/2` callback to
Adapters. In this callback they are responsible for converting a
Message into a Broadway.Message. The consumer takes care of running
the broadway apis so that the adapters don't need to.