Check for encoding errors in Google.Mock adapter (#43)
Protobuf schemas do not show any signs of errors until they are
encoded/decoded. This means you can create a protobuf struct and
basically pass anything you want to it.

```elixir
TestProto.new(lol: "oh yeah")
```

The current `Google.Mock` adapter does not do any encoding of the
message so you can send through a completely malformed protobuf and not
get any feedback there is an issue. Using the non mock Google adapter
would since it actually encodes the message as part of publishing.

This change makes it so that the `Google.Mock` adapter also runs
encoding on the message so that we can catch errors that we make in
testing and have more confidence in our messages.

RELATES: cu-vtbjvr