Set default concurrency for Broadway processors (#63)
We've been experiencing some flaky consumer tests in genesis-server that are
potentially due to the default concurrency option in Broadway starting up a
lot of processes on each test run; hoping that bringing down the default
processor concurrency will help. The default was System.schedulers_online() * 2.
Also we are only setting a default and this value can still be overridden if
needed.